Added: More emphasis on <th> inside <thead>
Since the fallback font Arial does not have a sensible font-weight: 500 and the Helvetica Neue looked too bold with 600, there's now added emphasis when using <thead>.

This is because the <th> can also be somewhere else, e.g. the first element of the row:
```html
<table>
  <tbody>
    <tr>
      <th>Name</th>
      <td>Cat</td>
      <td>Human</td>
    </tr>
    <tr>
      <th>Legs</th>
      <td>4</td>
      <td>2</td>
    </tr>
  </tbody>
</table>
```

In that case it makes no sense to always have the <th> tag have different margins and borders, just when it is inside the <thead> tag.